Migrate JsEval and SizeMeasure processors (#1598)(#1599)
