+++ streampipes-extensions/streampipes-processors-enricher-jvm/src/main/java/org/apache/streampipes/processors/enricher/jvm/processor/jseval/JSEvalProcessor.java
@@ -18,10 +18,21 @@
 package org.apache.streampipes.processors.enricher.jvm.processor.jseval;
 
 import org.apache.streampipes.commons.exceptions.SpRuntimeException;
+import org.apache.streampipes.model.DataProcessorType;
+import org.apache.streampipes.model.graph.DataProcessorDescription;
 import org.apache.streampipes.model.runtime.Event;
+import org.apache.streampipes.sdk.builder.ProcessingElementBuilder;
+import org.apache.streampipes.sdk.builder.StreamRequirementsBuilder;
+import org.apache.streampipes.sdk.helpers.CodeLanguage;
+import org.apache.streampipes.sdk.helpers.EpRequirements;
+import org.apache.streampipes.sdk.helpers.Labels;
+import org.apache.streampipes.sdk.helpers.Locales;
+import org.apache.streampipes.sdk.helpers.OutputStrategies;
+import org.apache.streampipes.sdk.utils.Assets;
 import org.apache.streampipes.wrapper.context.EventProcessorRuntimeContext;
 import org.apache.streampipes.wrapper.routing.SpOutputCollector;
-import org.apache.streampipes.wrapper.runtime.EventProcessor;
+import org.apache.streampipes.wrapper.standalone.ProcessorParams;
+import org.apache.streampipes.wrapper.standalone.StreamPipesDataProcessor;
 
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Value;
@@ -30,20 +41,39 @@ import org.graalvm.polyglot.proxy.ProxyObject;
 import java.util.HashMap;
 import java.util.Map;
 
-public class JSEval implements EventProcessor<JSEvalParameters> {
+public class JSEvalProcessor extends StreamPipesDataProcessor {
+
+  private static final String JS_FUNCTION = "jsFunction";
+
   private Context polyglot;
   private Value function;
 
   @Override
-  public void onInvocation(JSEvalParameters parameters, SpOutputCollector spOutputCollector,
+  public DataProcessorDescription declareModel() {
+    return ProcessingElementBuilder.create("org.apache.streampipes.processors.enricher.jvm.jseval")
+        .category(DataProcessorType.SCRIPTING)
+        .withAssets(Assets.DOCUMENTATION, Assets.ICON)
+        .withLocales(Locales.EN)
+        .requiredStream(StreamRequirementsBuilder
+            .create()
+            .requiredProperty(EpRequirements.anyProperty())
+            .build())
+        .requiredCodeblock(Labels.withId(JS_FUNCTION), CodeLanguage.Javascript)
+        .outputStrategy(OutputStrategies.userDefined())
+        .build();
+  }
+
+  @Override
+  public void onInvocation(ProcessorParams parameters,
+                           SpOutputCollector spOutputCollector,
                            EventProcessorRuntimeContext runtimeContext) throws SpRuntimeException {
     polyglot = Context.create();
-    String code = parameters.getCode();
+    String code = parameters.extractor().codeblockValue(JS_FUNCTION);
     function = polyglot.eval("js", "(" + code + ")");
   }
 
   @Override
-  public void onEvent(Event event, SpOutputCollector outputCollector) throws SpRuntimeException {
+  public void onEvent(Event event, SpOutputCollector collector) throws SpRuntimeException {
     // create new event with input event's source info and schema info.
     Event outEvent = new Event(new HashMap<>(), event.getSourceInfo(), event.getSchemaInfo());
     try {
@@ -54,7 +84,7 @@ public class JSEval implements EventProcessor<JSEvalParameters> {
       Map<String, Object> resultEvent = ((Value) result).as(java.util.Map.class);
       if (resultEvent != null) {
         resultEvent.forEach(outEvent::addField);
-        outputCollector.collect(outEvent);
+        collector.collect(outEvent);
       }
 
     } catch (ClassCastException e) {
@@ -63,8 +93,7 @@ public class JSEval implements EventProcessor<JSEvalParameters> {
   }
 
   @Override
-  public void onDetach() {
-    // do nothing.
-  }
+  public void onDetach() throws SpRuntimeException {
 
+  }
 }
